# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CareerPathr API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface StudentProfile {
  id: number;
  userId: number;
  yearOfStudy: string;
  major: string;
  skills: string[];
  interests: string[];
  updatedAt: string;
}

export interface StudentProfileInput {
  yearOfStudy: string;
  major: string;
  skills: string[];
  interests: string[];
}

export interface QuestionOption {
  value: string;
  label: string;
  careerTags: string[];
}

export interface Question {
  id: number;
  text: string;
  category: string;
  options: QuestionOption[];
}

export interface QuestionInput {
  text: string;
  category: string;
  options: QuestionOption[];
}

export interface AssessmentAnswer {
  questionId: number;
  selectedValue: string;
}

export interface AssessmentSubmission {
  userId: number;
  answers: AssessmentAnswer[];
}

export interface AssessmentResult {
  id: number;
  userId: number;
  completedAt: string;
  topCareerTags: string[];
  answers?: AssessmentAnswer[];
}

export interface Career {
  id: number;
  title: string;
  category: string;
  description: string;
  requiredSkills: string[];
  avgSalary: string;
  growthRate: string;
  learningTimeMonths: number;
}

export type CourseType = typeof CourseType[keyof typeof CourseType];


export const CourseType = {
  course: 'course',
  certification: 'certification',
  roadmap: 'roadmap',
} as const;

export type CourseLevel = typeof CourseLevel[keyof typeof CourseLevel];


export const CourseLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Course {
  id: number;
  careerId: number;
  title: string;
  provider: string;
  url: string;
  type: CourseType;
  duration: string;
  level: CourseLevel;
}

export interface CareerDetail {
  id: number;
  title: string;
  category: string;
  description: string;
  requiredSkills: string[];
  avgSalary: string;
  growthRate: string;
  learningTimeMonths: number;
  jobRoles: string[];
  futureOpportunities: string;
  courses?: Course[];
}

export interface CareerInput {
  title: string;
  category: string;
  description: string;
  requiredSkills: string[];
  avgSalary: string;
  growthRate: string;
  learningTimeMonths: number;
  jobRoles: string[];
  futureOpportunities: string;
}

export interface Recommendation {
  id: number;
  userId: number;
  careerId: number;
  matchScore: number;
  career: Career;
}

export interface SkillGapAnalysis {
  careerId: number;
  careerTitle: string;
  requiredSkills: string[];
  userSkills: string[];
  missingSkills: string[];
  matchedSkills: string[];
  matchPercentage: number;
}

export type CourseInputType = typeof CourseInputType[keyof typeof CourseInputType];


export const CourseInputType = {
  course: 'course',
  certification: 'certification',
  roadmap: 'roadmap',
} as const;

export type CourseInputLevel = typeof CourseInputLevel[keyof typeof CourseInputLevel];


export const CourseInputLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface CourseInput {
  careerId: number;
  title: string;
  provider: string;
  url: string;
  type: CourseInputType;
  duration: string;
  level: CourseInputLevel;
}

export type AdminStatsTopCareersByInterestItem = {
  careerTitle: string;
  count: number;
};

export interface AdminStats {
  totalStudents: number;
  totalCareers: number;
  totalQuestions: number;
  totalCourses: number;
  assessmentsCompleted: number;
  topCareersByInterest?: AdminStatsTopCareersByInterestItem[];
}

export type ListCareersParams = {
category?: string;
};

export type ListCoursesParams = {
careerId?: number;
};