# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CareerPathr API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get currently logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── STUDENT PROFILE ─────────────────────────────────────────────────────
...
  /assessment/submit:
    post:
      operationId: submitAssessment
      tags: [assessment]
      summary: Submit assessment answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentSubmission"
      responses:
        "200":
          description: Assessment processed, recommendations returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assessment/results/{userId}:
    get:
      operationId: getAssessmentResults
      tags: [assessment]
      summary: Get a user's latest assessment results
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessment results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResult"
        "404":
          description: No results yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── CAREERS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # ─── STUDENT PROFILE ────────────────────────────────────────────────────
...
    AssessmentAnswer:
      type: object
      required: [questionId, selectedValue]
      properties:
        questionId:
          type: integer
        selectedValue:
          type: string
    AssessmentSubmission:
      type: object
      required: [userId, answers]
      properties:
        userId:
          type: integer
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentAnswer"
    AssessmentResult:
      type: object
      required: [id, userId, completedAt, topCareerTags]
      properties:
        id:
          type: integer
        userId:
          type: integer
        completedAt:
          type: string
          format: date-time
        topCareerTags:
          type: array
          items:
            type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentAnswer"

    # ─── CAREER ─────────────────────────────────────────────────────────────
...