# lib/api-zod/src/generated/types/assessmentSubmission.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CareerPathr API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentAnswer } from './assessmentAnswer';

export interface AssessmentSubmission {
  userId: number;
  answers: AssessmentAnswer[];
}